requirements.txt
--only-binary=:all:
mujoco
numpy
pillow
pybind11-stubgen
matplotlib
imageio
imageio[ffmpeg] #both?